add: emotion bar
